requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.0
httpx~=0.25.2
Flask>=2.2.0
flask-cors>=3.0.10

